Integer overflow bug in low_cut_filter.
A multiplication result doesn't fit in an int32_t type. This change
converts one of the factors to int64_t to avoid overflows.

Bug: chromium:747202

+++ webrtc/modules/audio_processing/low_cut_filter.cc
@@ -59,7 +59,9 @@ class LowCutFilter::BiquadFilter {
       y[3] = y[1];
       y[0] = static_cast<int16_t>(tmp_int32 >> 13);
       y[1] = static_cast<int16_t>(
-          (tmp_int32 - (static_cast<int32_t>(y[0]) * ( 1 << 13))) * 4);
+          static_cast<int64_t>(tmp_int32 -
+                               (static_cast<int32_t>(y[0]) * (1 << 13))) *
+          4);
 
       // Rounding in Q12, i.e. add 2^11.
       tmp_int32 += 2048;
